modularize html structure and clean up imports
